Add `DeclarativeMetadata` to models
Trigger metadata refreshes when models change that feed into the metadata
attribute values.

Casting a wide net initially, if we find that we are issueing too many
refreshes we can be more granular with the refreshing (off specific attribute
changes).

We don't need to worry about over-refreshing metadata resulting in the
`api_updated_at` being touched as the service only touches the underlying data
if it has changed.