Remove `with_metadata` factory traits
We no longer need these as the metadata is created via callbacks.

You may run into an issue where the metadata is not created if you create the
object under test before the contract period and/or lead provider as we
currently do not have callbacks in these models. In this scenario you either
need to be explicit in the order you create your fixtures or call
`Metadata::Manager.refresh_all_metadata!` in a `before` block.