Prevent metadata from being created during data migration
We want to stop metadata from being created during a migration as we refresh
all the metadata at the end (which is more efficient/happens in a background
job).

Add a method to `Metadata::Manager` that accepts a block; for code executed in
the block no metadata will be created.

Wrap the contents of the `migrate` method in the `Migrators::Base` to ensure no
metadata is created.