Don't allow impersonating users back to admin
We could get into some weird situations if the admin is impersonating an
school user and navigates back to the admin section while impersonating.

To stop this from happening this change makes the
DfEUserImpersonatingSchoolUser no longer identify as DfEUser, just a
SchoolUser - which restricts them from visiting any non-school pages.

To get back to the admin section they need to stop impersonating, or
sign out and back in.