Rename current relationships to current_or_next
The behaviour of the #current_* relationships have morphed a little, and
was unclear, leading to misuse.

Current sounds like it means 'ongoing now', but we were using it for
'currently visible by the school' - hoping the person reading would know
that schools can see both the teachers who are currently there AND the
ones who are scheduled to join in the future.

Renaming to #current_or_next hopefully fixes this and makes the proc in
the has_one match the name of the relationship better.