School's can change an ECT's working pattern
Similarly to #1264, this allows school users to update ECT's working pattern.

This is pretty much the same implementation for changing an ECT's email address,
except in this journey we ask the user to press a button to change working
pattern (rather than submitting a form with a radio).

Even though we're using a hidden field, the `EditStep` does some basic
validations to ensure we don't "update" the working pattern with bad data. In
practice, though, these unhappy paths should never be travelled.