Don't reset the ECT wizard when navigating "back"
Before, we reset the ECT wizard every time a User visited the `/edit` step.

It turns out this wasn't the intention! Instead, design want the User's edit to
be "remembered" when they navigate back from the `/check-answers` page.

Now, in the `before_action` callback that resets the wizard, we check whether
the referer is the `check_answers` step. If it is (i.e the User clicked the
"Back" link), then we **don't** reset the wizard and the User sees the value
they previously entered.