Fix flaky ECT specs
Before, we always relied on the ongoing ECT and Mentor periods being generated
in the past.

But since the `started_on` dates are generated with a sequence, eventually we
start building periods that start in the future.

This meant we ran into test failures where date ranges were no longer
overlapping, or the earliest possible start date was in the future.

This updates the offending specs to initialise periods with explicit start
dates, so now they should **always** pass.