Ensure `ParityCheck::Client` makes consistent requests
We ran into an issue with the parity check where the request bodies were being
uniquely generated for a particular `ParityCheck::Request` (so one sent to ECF
and another to RECT).

This is hard to debug in the user interface; we have since fixed the underlying
issue, but this puts a safety net in place to ensure at the `Client` level it
is verifying the requests are being consistently made by comparing `method`,
`request_body`, `request_headers` and `query`.