Retain request URIs and body in `ParityCheck::Response`
We want to surface the request body and URIs in the parity check interface so
that we can easily double check the requests that were made.

Add request URIs to the `Response` model; we need to store both separately as
on occasion the paths will differ.

Add request body to the `Response` model; we only need to persist this once
for both ECF and RECT as we have tests that ensure the same request body is
used for both requests.