Introduce `ECTAtSchoolPeriods::SwitchTraining` service object
This was pulled out as part of the work to facilitate schools changing an ECT's
training programme.

There is some relatively complex business logic to consider when changing an
ECT's training programme from school-led to provider-led (and vice versa), so it
seemed sensible to abstract this into a service object we can test rigorously in
isolation.

This has been designed and implemented based on point 6. of the acceptance
criteria outlined [here]. Lots of this is new ground for me, so it's possible I
might have missed something!

The intention is to call this service object as part of the "Check answers" step
in the new "Change training programme" wizard, which will follow in a subsequent
PR.

[here]: https://github.com/DFE-Digital/register-ects-project-board/issues/2100