Schools can change an ECT's training programme
Similarly to other "Change" journeys, this allows schools to change an ECT's
training programme.

Unlike other journeys, this introduces some branching logic and an extra step.

"Switching" an ECT's training from provider-led to school-led (or vice versa) is
quite involved, so we can lean on the `SwitchTraining` service to handle these
changes.

This builds on work in #1382.

Users are presented with a straightforward choice when changing training
programme (i.e a CTA to confirm change, or not).

If switching from school-led to provider-led, the User is asked to choose a lead
provider in a follow up question.

Any change is made only after the check answers step has been completed.

Once the change is made, we record a new `teacher_training_programme_updated`
event.