Let periods transition on the same day
Before, a period couldn't finish on the same day it started.

This loosens that validation so we can switch training periods on the day they
started.

In practice, I'm not sure how often this will happen. But it seems reasonable
that a User might want to correct an error right away (which isn't possible
without this).

Now, a period can finish on the same day it started.