Add ECT/mentor became eligible attributes to `Teacher`
We want to track when an ECT or mentor first became eligible for training, as
we require this for the API. Once they have become eligible as far as the API
is concerned they won't revert.

Add attributes to `Teacher` to track first eligible dates.

Add validator to make dates immutable once set.