Let training periods switch before an ECT has started
This came out of product review when trying to change the training programme for
an ECT that hasn't started yet.

If we transitioned on the date of the change, it meant the old period would
finish before it had started (which led to validation errors or PG data errors).

If we transition on the ECT start date, we end up with multiple "current of
future" periods, which means the current one might be the "wrong" old one.

There is some more [context] on Teams.

To resolve this, we decided it makes most sense to destroy the existing training
period if the ECT hasn't started at the school yet.

This updates the service to accommodate that, plus adds a test to tell some of
the story.

[context]: https://teams.microsoft.com/l/message/19:wrVA9ZipsSjXLCrfIhG_aeXg61D4MjtXfyz2AR2lDjw1@thread.tacv2/1758908028470?tenantId=fad277c9-c60a-4da1-b5f3-b3b8b34a82f9&groupId=5e035efe-5b2b-491b-9e3e-d832445e4ad1&parentMessageId=1758908028470&teamName=Schools%20Digital&channelName=CPD%20Register%20ECTs&createdTime=1758908028470&ngc=true&allowXTenantAccess=true