Disable test guidance in the test env
We use the test guidance functionality to help with manual testing by
providing users with clues on how to fill in certain fields.

We never want this when running the automated tests as it can trigger
false positives - especially now as it controls the session debug info
in the footer. i.e. some tests were failing locally because '2025'
was expected on the page and also appeared in the session info.