Create training periods with consistent schools in specs
Before, it was possible for us to create `TrainingPeriod` records for ECTs or
Mentors where the `School` associated to the ECT or Mentor period was
**different** to the school associated to the training period via the school
partnership.

This doesn't reflect what happens in reality. If an ECT or Mentor is undergoing
provider-led training, this will be happening via a school partnership between
the lead provider and their school.

This tweaks the factory so we can pass through the school prevent this mismatch.