Introduce `Schools::ECTs::LeadProviders::Assignable` concern
We want to assign a lead provider in the "Change mentor" and "Change training
programme" journey.

Since they are sharing a lot of common logic, it made sense to abstract away
this methods in a concern. This pattern felt reasonable, since we're making use
of concerns in the "Change" wizard journeys already.

There shouldn't be any change in behaviour as a result of this change.