Schools can change an ECT's mentor to another registered mentor
This builds on some similar journeys and introduces a "Change mentor" journey
for schools.

Even though this might seem straightforward, there is a bunch of logic and a few
different branches to cover.

This change covers

- Changing to a registered mentor who is not eligible for training
- Changing to a registered mentor who is eligible for training and using the
  existing provider
- Changing to a registered mentor who is eligible for training and choosing a
  different provider

Lots of this is tests and boilerplate, and some steps have been taken to reuse
some aspects that are similar in other journeys.

There is some more work to do in this area, but clear abstractions haven't
presented themselves just yet. There is some work in the "Assign mentor" journey
we want to reuse, but this is a decent starting point.