Enable schools interface in tests with a flag
Before, we needed to stub the `enable_schools_interface` configuration
explicitly in every test.

As we get closer to launching the schools interface, and build out more of the
functionality, we need to do this more and more.

This stubs the `enable_schools_interface` configuration based on the presence of
an `:enable_schools_interface` flag in the example metadata.