Add `Teachers::Query`
Add the `Teachers::Query` service to allow us to query teachers accoording
to the API specification.

Note that we have to `eager_load` instead of `includes` when preloading
the associations in order to be able to filter out the metadata not
associated with the given lead provider. As `eager_load` can be a little
less efficient that `includes` (it creates a single, bloated query and the
results are paired back in memory) we may want to drop the metadata
filtering and `includes` instead if we find performance is an issue.

Filtering on `contract_period_year` is also non-trivial in this query due
to needing to filter across two relationships
(`latest_ect_training_period` OR `latest_mentor_training_period`). The
alias names for joined tables Rails generates are not very clear. We will
look into adding the contract periods to `Metadata::TeacherLeadProvider`
to clean this up and make it more efficient in a follow up ticket.