Prevent implicit teachers from being seeded
Currently, after running the seeds we have a small number of teachers that are
created implicitly by the factories and are not defined in the `teachers.rb`
seed.

This happens because the `teacher_id_change` factory will always create a new
teacher to assign the `api_from_teacher_id` to. Instead, we can make this
create a new teacher only if one is not specified and explicitly pass in the
`api_from_teacher_id` in the teacher seeds.