Refresh `Metadata::TeacherLeadProvider` for `api_mentor_id`
We need to refresh the `Metadata::TeacherLeadProvider` to update
`mentor_api_id` when any of the following occur:

- `ECTAtSchoolPeriod` created/destroyed
- `MentorshipPeriod` created/destroyed

This assumes the `mentor` will not change on a `MentorshipPeriod`.

The `teacher` metadata is already refreshed when a `TrainingPeriod` is
created/destroyed/updated.

Minor update to the declarative updates shared context as we now have scenarios
where the `instance` and `target` refresh the metadata on `create`, so we have
multiple calls to `refresh_metadata!`.