Fix flaky test
As we add more tests, the base dates generated in various factories can get
bigger and bigger.

My hypothesis is this is why we are seeing more order-dependent flaky tests.

This explicitly sets the `ECTAtSchoolPeriod` started on date so it is inside the
relevant `ContractPeriod` started on date. Now, the tests always pass.