Schools can change an ECT's lead provider
This introduces a new "Change" journey, which allows Schools to change an ECT's
lead provider.

This is a basic journey and it leans on the `TrainingPeriodSources` concern to
ensure the relevant training period relationships are created.

Similarly to the "Change training programme" journey, there is some simple logic
to ensure we destroy training periods that have not begun or do not have
confirmed partnerships with schools.