Implement GET participants show/index actions
Fix up the routes so the correct `api_id` parameter is used (as it was
previously the id parameter is inferred as `participants_api_id`).

Fix eager loading in the `Participants::Query` and use `include` on
`teacher_id_changes` for better performance.

Implement `ParticipantsController`, calling the existing query and
serializer.

Note that the for consistency with other endpoints and to make the
implementation more straight forward we can accept multiple contract
periods when filtering. We will need to reflect this in the API
documentation but it should be a backwards compatible change.