Add `training_status` filtering to participants endpoint
We need to allow lead providers to filter participants by a given
`training_status`. As we don't store the training status like we did in
ECF we have to calculate this state on the fly.

We have a service to calculate it for a given `TrainingPeriod`, but in
the query we need to replicate the logic to determine the state
efficiently.

We surface the teacher if either of their training periods match the
given state.