Improve teachers serializer and preloading
The serializer was walking back up through `trainee#teacher` to access
associations back on the base teacher, which would then require their own
preloading. Instead, we can pass the association serializer the
`TrainingPeriod` and the `Teacher` so that we have the higher level `Teacher`
available thats already preloaded.

This means we can shift a bunch of the eager loads onto the parent `Teacher`
using the `includes` which should be more efficient.