Add trnless flag and update TRN index
We have some historic cases where some ECF1 records have no TRN, but we
have to import them.

For all new records we want TRNs to be present, and ideally we want to
keep the database level checks to ensure we don't accidentally add bad
data.

This change does the following:

1. it adds a trnless boolean field to teachers
2. when trnless: true
  - the validation in the model skips the TRN/uniqueness check
  - the database level validation for TRN presence isn't applied
3. when trnless: false (default)
  - the validation in the model is applied
  - the database level validation is applied