Optimise `Teachers::Query`
We were using `eager_load` here in order to filter joined metadata to
only the lead provider we care about, however this resutls in a lot of
`LEFT OUTER JOINS` and a large/inefficient query.

Switch to `include` to improve the query and drop the filtering by lead
provider as we can't do this with `includes` here and the trade off is
worth it.

Oddly in other query services we can filter by LP metadata with
`include`; it seems to be a side effect of this more complicated series
of joins causing it to not work, but I haven't dug into the actual
issue further.