Schools can change an ECT's mentor to another registered mentor
This builds on some similar journeys and introduces a "Change mentor" journey
for schools.

Even though this might seem straightforward, there is a bunch of logic and a few
different branches to cover.

This change covers:

- Changing to a registered mentor who is not eligible for training
- Changing to a registered mentor who is eligible for training and using the
  existing provider
- Changing to a registered mentor who is eligible for training and choosing a
  different provider
- Changing to a registered mentor when the ECT is undergoing school-led training

Lots of this is tests and boilerplate.

Some steps share quite a few methods. Previous iterations of this change
experimented with a few different abstractions, but none quite felt right.

This keeps things explicit and verbose (if a little repetitive) for now. We can
always circle back later and make improvements if things become clearer.