WIP - Changing mentor multiple times on the same day
Failing test due to unique constraint for mentorship periods for a given
`ect_at_school_period_id` and `started_on`.

Are we happy to ship with this bug and revisit?

Or do we want to loosen the database constraint? (e.g exclude mentorship periods
with a `finished_on`)

Or something else?