Schools can change an ECT's lead provider
This introduces a new "Change" journey, which allows Schools to change an ECT's
lead provider.

This is a straightforward journey, but there is still considerable business
logic to consider when switching from one lead provider to another.

For now, this is abstracted away in the `ECTAtSchoolPeriods::SwitchLeadProvider`
service object. But it's possible we might unearth something more useful as we
continue to build out these "Change" journeys.