Return no results when filtering on only invalid contract periods
To be consistent with ECF we want to return no results if only an
invalid cohort is provided when filtering. At the moment, this will be
the case if the cohort is numerical, so `1999` or `0000`, however
non-numerical cohorts are stripped out via `extract_conditions(...,
integer: true)`.

This means we end up passing an empty array into the query, which is
ignored via `ignore?(filter:)`.

To get around this, we can now pass `ignore_empty_array: false` to
`ignore?(filter:)` and use this in the query services to join on an
empty array which will return 0 results and is consistent with ECF.

Extra tests added to shared request specs to ensure consistency.