Add Gantt charts to the migration teacher page
These charts will help us get a quick overview of a teacher's induction
records in ECF1. It is useful because sometimes they're hard to follow
because the teacher changes school, lead provider, training programme,
mentor etc can change.

I had to adjust the CSV building a bit so we can still make use of the
struct, otherwise there's loads of traversing of AR records in the
consuming classes which is a bit ugly.

It allows the CSV writing to be tidied up a little too.