Add helper to build custom swagger response examples
By default swagger will build the examples from the schema files. On some
endpoints we want to tweak the response examples so they may more sense to the
context of the endpoint (setting a field the endpoint might populate, for
example).

Add `SwaggerExampleParser` to easily load the default swagger response example.

Add `after` hooks into shared contexts so that we can set custom response
examples.

Set `dry-run` to `false` for running `swaggerize`; we need to do this to have
it call the `after` blocks that we use to customise the response.

Note that we have intentionally opted not to populate the responses with the
response bodies of the specs, as our specs are random and we don't want lead
providers to have random data in the documentation. We also don't want to
couple our factory data to our swagger documentation -- all examples should be
hard coded in the schema files.