Fix duplication in schools `filter[cohort]`
When `dry-run` is enabled in `rswag` it executes the underlying tests. We have
a workaround to pass the required `filter[cohort]` value in the schools request
which involves appending a new query parameter. This now results in duplication
in the swagger file output, so we need to manually remove it after running the
example.