Include school partnerships in API expression of interests
When we return the `expression_of_interest` attribute in the API we need to
also set this to `true` if there is a school partnership. This is because we
will not be populating `expression_of_interest` for historical `TrainingPeriod`
records imported from ECF, yet we need to indicate there was an expression of
interest to lead providers.

Update `School` handler to include school partnerships in expression of
interest calculation. Memoize and optimise the query to calculate for all LP/CP
combinations in a single query.