Optimise `Schools::TrainingProgramme` service
Previously, we instantiated a new service for each training programme
calculation/contract period year. This doesn't scale well when we want to know
the training programme for all contract period years (as we do for calculating
metadata).

Instead, we can create a single instance for a school and query across all
contract periods. This allows us to efficiently calculate training periods for
all contract periods relative to a school.