Fix swagger documentation for participants/errors
We had the `ecf_enrolments` attribute listed as `enrolments` in the swagger
schema, however we are still using `ecf_enrolments` in the serializer (and want
to continue to for now).

This was missed because rswag ignores extra response keys by default.

Switch rswag to raise on extra response keys.

Rename `enrolments` to `ecf_enrolments` to match serializer.

Correct 404 response schema to match what we actully return/what ECF uses.