Add indexes to improve `Teachers::Query` performance
The teachers query is currently quite slow (~1.5s response time).

Our query is fairly optimal as it stands, but we're lacking indexes in a couple
of places to avoid full table scans and to speed it up.

Add a minimal amount of indexes; in testing this brings the response time down
to 5-600ms, which I think is faster than ECF so we can stop there for now.

There is an upcoming PR to stamp contract period on the metadata which will
improve that query/response time.