Withdraw participant API documentation
This adds swagger documentation for the PUT participant `/withdraw` API endpoint.

The example response has been overriden to return a participant with a training
status of "withdrawn", no deferral and a withdrawal.

As part of this change, we have introduced a helper method that overrides the
response content for a given example. This is reused across shared contexts in
the API docs specs.