Refresh metadata when school partnership is set
We hit a bug during PO review whereby when we create a `SchoolPartnership` and
it is later assigned to a `TrainingPeriod`, the `Metadata::TeacherLeadProvider`
metadata was not being refreshed.

Update to refresh when `school_partnership_id` is changed to cover this
scenario.

Update school metadata refresh to only happen when `school_partnership_id` or
`expression_of_interest_id` is changed; this should avoid unnecessary updates.