Resume participants PUT endpoint
Similarly to #1610 and #1597, this implements the `/resume` participants API
endpoint.

Now, lead providers can make a PUT request to the API endpoint to resume a
participant's training.

Since #1638, the `course_identifier` param given by providers is extracted to
the `teacher_type` by the `API::TeacherType` concern.

For this controller action, there are no other params. This means we don't
respond with `400 Bad Request` when the incorrect JSON structure is used in the
request body. Now, we need to validate the controller params explicitly.