Fix foreign key values in declarative update specs
Previously, we generated a `new_value` just based on the `integer` column type
via `rb_integer`, however this can lead to generating ID values that already
exist in the database.

Instead, we can use reflection to find the type of model for the association
and query the next `id` in the sequence from the database, ensuring we'll
always be able to create the associated model.