Improve parity check match rate calculation
Previously, we would treat a perfect match as `100` and a response with any
difference (in status code or response body) as `0`. This is quite misleading,
as often response bodies _mostly_ match yet the `0` bubbles up and the overall
run success is very low.

Instead, we can calculate a reasonable match rate by taking the number of
different lines relative to the total lines of the JSON responses. Again, this
won't be perfect, but should be a better indication of how different the
responses actually are.

The match rate currently takes into account the status code comparison, which
makes it a bit awkward but we can just fall back to a `0` match rate if the
status codes are different (generally speaking if the status codes are
different the bodies would be completely different anyway, so it seems
reasonable to do).

We store the calculated match rate as running the `diff` is reasonably
expensive and adds up to a lot of time when calculating it for an overall `Run`
(as it needs to diff every `Response`).