Support excluding attributes in parity check
For some responses we expect there to be extra or missing attributes; we can
safely exclude these from the parity check results.

In order to do this we need to access the request -> endpoint -> options when
persisting the response, so this commit reworks to assign the request in the
`Client` so that we have it in time.

We also rework the `advance_page` to ensure its called prior to the `Response`
being persisted, as we `nil` out equal response bodies to save DB space when
they are equal (and therefore lose the information we need to determine if
there is another page).