Ensure participants cannot be resumed with ongoing training
Before, it was possible for a participant to be resumed even when they had an
ongoing training period at another lead provider.

This shouldn't be possible, because their training status for that lead provider
is "active".

This change adds an extra validation that checks whether there are any other
training periods that are "ongoing today" associated with the school period.

Now, when a participant is undergoing training at another lead provider they can
no longer be "resumed" by another lead provider where their latest training
period at that provider is "withdrawn" or "deferred".