Add `api_updated_at` to training periods
We want to return an `updated_at` for "Tranfers" in the API.

Roughly speaking, a transfer represents the transition of a Teacher's training
period **between** schools.

A transfer always has a "leaving" training period, and sometimes has a "joining"
training period.

Since we don't model a transfer directly, we need to determine the time a
transfer was last updated by looking at the relevant training periods.

This introduces a new column to training periods so we can store the
`api_updated_at` timestamp. This follows a similar pattern to Schools, Teachers
etc.