Tweak LP currently training validation
Talking with Claire and Andy its going to be enough to make sure that the
training period is 'ongoing today'. We will allow the current lead provider to
change schedule even if they are moving to a new lead provider in the future.

We want to ensure that if they are moving to a new school or LP in the future
that we cap the newly created training period `finished_on` appropriately to
avoid overlaps.