Change schedule behaviour changes
We only want to allow a lead provider to change schedule if they are currently
training the participant **and no future training periods exist**. This is due
to future training periods potentially being incompatible with the contract
period they are changing to (if a `SchoolPartnership` does not exist for the
future lead provider, for example).

We may revisit this in the future as there will be a period of time when a
participant is scheduled to start at a new school with a new LP where neither
LP will be able to change their schedule, but we would expect this to be an
edge case.

We also need to determine the correct `finished_on` behaviour for the existing
and new training period. If the existing training period has finished already,
we retain that `finished_on`. If it finishes in the future we can truncate the
`finished_on` to today in order to be able to slot the new training period in.
The new training period will match the `finished_on` of the training period we
are closing.