Don't retain `expression_of_interest` across training periods
When a new training period is created we don't want to bring across the
`expression_of_interest`. We surface this in the API if _any_ training period
has it set relative to a school and it only makes sense to set it on the
training period that it was originally actioned on.