WIP - Transfer history for a given Teacher and Lead Provider
We want to compile all the transfers for a given Teacher relative to a lead
provider, so we can return a list of transfers in the API.

Each transfer has a leaving `TrainingPeriod` and an optional joining
`TrainingPeriod`. There are some rules that dictate what status and type a
transfer is, and these are documented in the tests.

The tests need to be extended to test mentor transfers too, and will likely be
reworked to better match the style of tests we use throughout the application.