Disable `RSpec/InstanceVariable` for the transfer history specs
We decided to write the tests using `before` blocks and using instance
variables, rather than leaning on `let`.

This decision was not made lightly! After plenty of back and forth, we felt this
was more readable and painted a clearer picture of the relationships between the
different school periods, training periods and lead providers.