Remove default independent_school trait
The :ect_at_school_period_factory defaulted to the :independent_school
type which also creates an AB via the :national_ab trait.

Now we're using factories for seeds we don't want them, so default to
just creating a default school.

This takes the number of ABs generated by the seeds down from 100+ to
10.