Add `involved_in_school_transfer` to `TeacherLeadProvider` metadata
As we pivot away from the approach outlined in #1760 to one that uses metadata,
we want to introduce a boolean attribute to the `Metadata::TeacherLeadProvider`
model.

This will make the query to retrieve all the Teachers with a school transfer for
a given Lead Provider much more straightforward.

This column is nullable for now, but we might consider making it non-nullable
once we have refreshed the metadata for all Teachers.