Refresh induction period teacher metadata
Whether or not a Teacher is involved in a school transfer for a given Lead
Provider depends partly on whether they have completed their induction.

This means we need to refresh a Teacher's metadata when an `InductionPeriod` is
completed (i.e `finished_on` and `outcome` are not nil).

This uses declarative updates to refresh the metadata when either of those
attributes change.

In order to make this change, we needed to update the `InductionPeriod` model to
define an enum for `outcome` to align it with the DB column type.