Fix flaky `ActiveLeadProvider` spec
This is failing intermittently because `different_year_alp` used a default contract period sequence. The contract period is created as 2021 then incrementally increases eventually reaching 2025 which is the contract period that we're intentionally trying to avoid.

The fix is to use a relative contract period year to the original contract period, that way they can never be the same.