Prevent zero-day training periods on participant actions
We want to avoid zero-day training periods when a lead provider
defers/withdraws and then resumes and defers/withdraws again in quick
succession (on the same day).

Currently, this would raise an error/exception as we would eventually end up
with a training period that is created/finished on the same day and we require
a minimum 1 day duration for all training periods.

We want to handle this by removing the zero-day, resumed training period, for
example:

- Provider defers a TP (this sets the finished_on)
- Provider resumes on the same date (this creates a new TP)
- Provider then defers again:
  - Delete the latest, would-be zero-day TP
  - Set deferred reason/date

If there is no previous training period to 'rollback' to, we will raise an
error as we can't gracefully handle the scenario.

As part of this change we are now allowing a training period to be both
withdrawn and deferred. We will always take the latest as their training status
and only ever surface the latest in the API. We should have allowed this
initially anyway as moving a participant from deferred to withdrawn is a valid
transition.

Adds a test to ensure a participant can go from withdrawn to deferred.

In the rare case that a lead provider attempts to withdraw/defer a participant
on the very first day they start training with them we will return an error (as
if we finish the TP it would be zero-day and we have nothing to fallback to).