Introduce `API::Teachers::SchoolTransfers::Query`
This query object returns all the participants with school transfers for a given
lead provider.

It leans on established patterns used in other query objects in the API.

Since the query to retrieve relevant Teachers proved to be extremely complex
(see #1760), we decided to use a flag on the `TeacherLeadProvider` metadata
instead.

This makes the query much more straightforward, and means all the logic around
what is/isn't a "school transfer" is contained in the
`Teachers::SchoolTransfers::History` class.