Add `api_transfer_updated_at` to `TrainingPeriod`
We need a reliable way of determining when a transfer surfaced in the API has been touched/changed. A transfer is a `TrainingPeriod` that is leaving/joining a school (excluding the first training period and last training period for participants that have completed training).

Add `api_transfer_updated_at` to `TrainingPeriod`.

Touch the new field when any of the following are changed:

- School#urn
- TrainingPeriod#started_on
- TrainingPeriod#finished_on
- LeadProvider#name

Update the query and serializer to work with the new updated at attribute.

Remove tests around sorting by `updated_at` in the query service; we don't
support custom ordering in the transfers endpoint (and sorting on
`api_transfer_updated_at` leads to a very complex query, so I don't think we
should add it).