Refactor `Teachers::SchoolTransfers::History`
In the serializer we will have access to the lead provider ID rather than the
lead provider record, but until now the `History` compared lead provider
records.

This refactors the `relevant_to_lead_provider?` check to compare IDs, and
reworks the initializer so it accepts IDs instead of lead providers.

As part of this change we also added tests to ensure the `History` class can
handle scenarios where there are no transfers or a teacher has no training
periods.