Modelling for declaration and statement items
Finishes the `Declaration` model by adding the missing associations
and attributes from ECF.

Adds two new models; `Statement::PaymentItem` and
`Statement::ClawbackItem`. This diverges from ECF, which has a
single `StatementItem`, however it hopefully makes the intention and
behaviour of the models clearer by splitting them up.

A `Declaration` moves through a number of states; when `submitted`
it is neither payable or a clawback; but once it changes state it
can move through payable and clawback states.

A `Declaration` can only ever have one `PaymentItem` and one
`ClawbackItem` (or neither if it remains in the `submitted` state).