Finish `Declaration` modelling
The `Declaration` model has been stubbed out up until now; we need to flesh
this out ahead of the declarations work for the API.

Add missing attributes to `Declaration` and change `declaration_type` to an
`enum` to match what we use on `Milestone`.

We are adding a couple of new relationships; `voided_by_user` to track which
admin user has manually voided a declarration and `mentor_id` which will stamp
the `Declaration` with the `Teacher` that was mentoring the participant at the
time it was made. We are stamping this as it is needed in the serializer and
would involve an expensive and complex set of joins to pull it dynamically (we
stamp it in ECF for this reason as well).

Add appropraite validation logic as per ECF.

Control state with a state machine to be consistent with `Statement` and
`Statement::LineItem`.